split demo into individual parts + DB provider works without DB

+++ client/src/db/DbTest.tsx
@@ -1,11 +1,11 @@
-import { useDB, useQuery } from "@vlcn.io/react"
 import { useEffect, useState } from "react"
 import { sql } from "shared/sql"
 
 import { useSync } from "client/db/Sync"
 import { useDbQuery } from "client/db/useDbQuery"
 import { useDbMutation } from "client/db/useDbMutation"
-import { Button } from "client/Button/Button"
+import { Button } from "client/components/Button/Button"
+import { useDB } from "client/db/useDb"
 
 function Test({ name }: { name: string }) {
 	const other = useDbQuery<{ id: string; content: string; position: number }>({
@@ -49,7 +49,7 @@ function Test2({ name }: { name: string }) {
 
 export function Content({ name }: { name: string }) {
 	const ctx = useDB(name)
-	const sync = useSync(ctx.db, name)
+	const sync = useSync(ctx?.db, name)
 
 	const { mutateAsync } = useDbMutation<[id: string, content: string]>({
 		dbName: name,
@@ -61,22 +61,23 @@ export function Content({ name }: { name: string }) {
 		await sync?.roundTrip()
 	}
 
-	const { mutateAsync: mutateAsync2 } = useDbMutation<
-		[id: string],
-		{ id: string; content: string }
-	>({
-		dbName: name,
-		query: sql`DELETE FROM test WHERE id = ? RETURNING id, content;`,
-		returning: true,
-	})
+	// const { mutateAsync: mutateAsync2 } = useDbMutation<
+	// 	[id: string],
+	// 	{ id: string; content: string }
+	// >({
+	// 	dbName: name,
+	// 	query: sql`DELETE FROM test WHERE id = ? RETURNING id, content;`,
+	// 	returning: true,
+	// })
 
-	const removeData = async (id: string) => {
-		const [res] = await mutateAsync2([id])
-		console.log("---- remove", res)
-		await sync?.roundTrip()
-	}
+	// const removeData = async (id: string) => {
+	// 	const [res] = await mutateAsync2([id])
+	// 	console.log("---- remove", res)
+	// 	await sync?.roundTrip()
+	// }
 
 	const dropData = async () => {
+		if (!ctx) return
 		await ctx.db.exec(sql`DELETE FROM test;`)
 		await sync?.roundTrip()
 	}
@@ -88,11 +89,6 @@ export function Content({ name }: { name: string }) {
 		if (navigator.onLine) sync.roundTrip()
 		return () => controller.abort()
 	}, [sync])
-
-	const result = useQuery<{ id: string; content: string }>(
-		ctx,
-		sql`SELECT id, content, position FROM test ORDER BY position, id ASC`
-	)
 
 	const [toggle, setToggle] = useState(true)
 	const [toggleBis, setToggleBis] = useState(false)
@@ -111,14 +107,6 @@ export function Content({ name }: { name: string }) {
 				Toggle 2 {String(!toggle2)}
 			</Button>
 			{toggle2 && <Test2 name={name} />}
-			<ul>
-				{result.data?.map((item) => (
-					<li key={item.id}>
-						{item.content} <Button onClick={() => removeData(item.id)}>delete</Button>
-					</li>
-				))}
-			</ul>
-			<hr />
 			<hr />
 			<form
 				onSubmit={(event) => {
